requirements.txt
tencentcloud-sdk-python
Crypto
pycryptodome
